@@ -1,9 +1,9 @@
 <?php
-/**
+/*
  * This file is part of Berlioz framework.
  *
  * @license   https://opensource.org/licenses/MIT MIT License
- * @copyright 2017 Ronan GIRON
+ * @copyright 2020 Ronan GIRON
  * @author    Ronan GIRON <https://github.com/ElGigi>
  *
  * For the full copyright and license information, please view the LICENSE
@@ -16,6 +16,7 @@ namespace Berlioz\Router;
 
 use Berlioz\Http\Message\Request;
 use Berlioz\Router\Exception\RoutingException;
+use Psr\Http\Message\ServerRequestInterface;
 
 use const ARRAY_FILTER_USE_KEY;
 
@@ -23,144 +24,210 @@ use const ARRAY_FILTER_USE_KEY;
  * Class Route.
  *
  * @package Berlioz\Router
- * @see \Berlioz\Router\RouteInterface
  */
 class Route implements RouteInterface
 {
-    protected const REGEX_PARAMETER = '/{(?<name>[\w_]+)}/';
-    /** @var string Route */
-    private $route;
-    /** @var string[][] Route options */
-    private $options;
-    /** @var string Route regex */
-    private $route_regex;
-    /** @var array Context */
-    private $context;
-    /** @var Parameter[] Parameters */
-    private $parameters;
+    use RouteSetTrait;
+
+    protected const REGEX_ATTRIBUTE = '/{(?<name>' . Route::REGEX_ATTRIBUTE_NAME . ')(?:::(?<type>\w+)|:(?<regex>[^}]+))?}/i';
+    protected const REGEX_ATTRIBUTE_NAME = '[\w_]+';
+    protected const REGEX_ATTRIBUTE_VALUE = '[^/]+';
+    private ?array $method = null;
+    private ?array $host = null;
+    /** @var Attribute[] Attributes */
+    protected array $attributes = [];
+    protected ?Route $parent = null;
+    protected ?string $path_regex = null;
 
     /**
      * Route constructor.
      *
-     * @param string $route
+     * @param string|null $path
+     * @param array $defaults
+     * @param array $requirements
+     * @param string|null $name
+     * @param string|array|null $method
+     * @param string|array|null $host
+     * @param int $priority
      * @param array $options
-     * @param array $context
+     * @param mixed $context
      *
      * @throws RoutingException
      */
-    public function __construct(string $route, array $options = [], array $context = [])
-    {
-        $this->route = $route;
-        $this->options = $options;
-        $this->context = $context;
-        $this->parameters = [];
+    public function __construct(
+        protected ?string $path = null,
+        array $defaults = [],
+        array $requirements = [],
+        protected ?string $name = null,
+        string|array|null $method = null,
+        string|array|null $host = null,
+        protected int $priority = -1,
+        protected array $options = [],
+        protected mixed $context = null,
+    ) {
+        // Extract attributes from path
+        $this->path =
+            preg_replace_callback(
+                static::REGEX_ATTRIBUTE,
+                function ($matches) {
+                    $name = $matches['name'];
 
-        // Read parameters
-        $matchesParams = [];
-        if (preg_match_all(self::REGEX_PARAMETER, $this->route, $matchesParams) > 0) {
-            foreach ($matchesParams['name'] as $match) {
-                if (isset($this->parameters[$match])) {
-                    throw new RoutingException(sprintf('Duplicate attribute "%s" in route "%s"', $match, $route));
-                }
+                    // Duplicated attribute
+                    if (!array_key_exists($name, $this->attributes)) {
+                        $attribute = new Attribute($name, route: $this);
+                        $this->attributes[$name] = $attribute;
 
-                $parameter = new Parameter(
-                    $match,
-                    $this->options['defaults'][$match] ?? null,
-                    $this->options['requirements'][$match] ?? null
-                );
+                        if (!empty($matches['regex'])) {
+                            $attribute->setRegex($matches['regex']);
+                        }
+                        if (!empty($matches['type'])) {
+                            $this->attributes[$name]->setRegex(
+                                Attribute::TYPES[$matches['type']] ??
+                                throw new RoutingException(sprintf('Unknown type "%s"', $matches['type']))
+                            );
+                        }
+                    }
 
-                $this->parameters[$match] = $parameter;
-            }
+                    return '{' . $name . '}';
+                },
+                $this->path
+            );
+
+        // Define requirement attributes
+        foreach ($requirements as $name => $requirement) {
+            $attribute = $this->attributes[$name] ?? $this->attributes[$name] = new Attribute($name, route: $this);
+            $attribute->setRegex($requirement);
+        }
+
+        // Define default attributes
+        foreach ($defaults as $name => $default) {
+            $attribute = $this->attributes[$name] ?? $this->attributes[$name] = new Attribute($name, route: $this);
+            $attribute->setDefault($default);
+        }
+
+        if (null !== $method) {
+            $this->method = (array)$method;
+            array_walk($this->method, fn(&$value) => $value = strtoupper($value));
+        }
+
+        if (null !== $host) {
+            $this->host = (array)$host;
+            array_walk($this->host, fn(&$value) => $value = strtolower($value));
         }
     }
 
     /**
-     * @inheritdoc
+     * PHP serialize method.
+     *
+     * @return array
      */
-    public function serialize(): string
+    public function __serialize(): array
     {
-        return serialize(
-            [
-                'route' => $this->route,
-                'options' => $this->options,
-                'route_regex' => $this->route_regex,
-                'context' => $this->context,
-                'parameters' => $this->parameters,
-            ]
-        );
+        $this->compile();
+
+        return [
+            'path' => $this->path,
+            'name' => $this->name,
+            'method' => $this->method,
+            'host' => $this->host,
+            'priority' => $this->priority,
+            'attributes' => $this->attributes,
+            'options' => $this->options,
+            'context' => $this->context,
+            'path_regex' => $this->path_regex,
+        ];
     }
 
     /**
-     * @inheritdoc
+     * PHP unserialize method.
+     *
+     * @param array $data
      */
-    public function unserialize($serialized)
+    public function __unserialize(array $data): void
     {
-        $tmpUnserialized = unserialize($serialized);
+        $this->path = $data['path'];
+        $this->name = $data['name'];
+        $this->method = $data['method'];
+        $this->host = $data['host'];
+        $this->priority = $data['priority'];
+        $this->attributes = $data['attributes'];
+        $this->options = $data['options'];
+        $this->context = $data['context'];
+        $this->path_regex = $data['path_regex'];
 
-        $this->route = $tmpUnserialized['route'];
-        $this->options = $tmpUnserialized['options'];
-        $this->route_regex = $tmpUnserialized['route_regex'];
-        $this->context = $tmpUnserialized['context'];
-        $this->parameters = $tmpUnserialized['parameters'];
+        // Set group parent
+        array_map(fn(Route $route) => $route->setParent($this), $this->routes);
+        array_map(fn(Attribute $attribute) => $attribute->setRoute($this), $this->attributes);
     }
 
     /**
-     * @inheritdoc
+     * Get name.
+     *
+     * @return string|null
      */
-    public function getName(): string
+    public function getName(): ?string
     {
-        if (!empty($this->options['name']) && is_string($this->options['name'])) {
-            return (string)$this->options['name'];
+        return $this->name;
+    }
+
+    /**
+     * Get path.
+     *
+     * @return string
+     */
+    public function getPath(): string
+    {
+        $path = $this->parent?->getPath();
+
+        return (($path ?? '') . $this->path) ?? '';
+    }
+
+    /**
+     * Get path regex.
+     *
+     * @return string
+     */
+    protected function getPathRegex(): string
+    {
+        if (null === $this->path_regex) {
+            $this->compile();
         }
 
-        return spl_object_hash($this);
+        return $this->path_regex;
     }
 
     /**
-     * @inheritdoc
-     */
-    public function getOptions(): array
-    {
-        return $this->options ?? [];
-    }
-
-    /**
-     * Get option.
+     * Get attribute.
      *
      * @param string $name
-     * @param mixed $default
      *
-     * @return mixed
+     * @return Attribute|null
      */
-    public function getOption(string $name, $default = null)
+    public function getAttribute(string $name): ?Attribute
     {
-        if (array_key_exists($name, $this->options)) {
-            return $this->options[$name];
-        }
+        return $this->attributes[$name] ?? $this->parent?->getAttribute($name);
+//        $parentAttribute = $this->parent?->getAttribute($name);
 
-        return $default;
+        // If not exists, return parent attribute
+//        if (!array_key_exists($name, $this->attributes)) {
+//            return $parentAttribute;
+//        }
+
+//        $attribute = $this->attributes[$name];
+//        if (!$attribute->hasRegex()) {
+//            $attribute->setRegex($parentAttribute?->getRegex());
+//        }
+//        if (!$attribute->hasDefault()) {
+//            $attribute->setDefault($parentAttribute?->getDefault());
+//        }
+//
+//        return $attribute;
     }
 
     /**
-     * @inheritdoc
-     */
-    public function getContext(): array
-    {
-        return $this->context ?? [];
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function setContext(array $context): RouteInterface
-    {
-        $this->context = $context;
-
-        return $this;
-    }
-
-    /**
-     * @inheritdoc
+     * Get methods.
+     *
+     * @return array
      */
     public function getMethods(): array
     {
@@ -175,183 +242,294 @@ class Route implements RouteInterface
             Request::HTTP_METHOD_DELETE,
         ];
 
-        $methods = [];
-        if (isset($this->options['method'])) {
-            if (is_scalar($this->options['method'])) {
-                $methods = explode(',', (string)$this->options['method']);
-            } else {
-                $methods = (array)$this->options['method'];
+        if (null === $this->method) {
+            if ($this->parent) {
+                return $this->parent->getMethods();
             }
-            $methods = array_map('mb_strtoupper', $methods);
-            $methods = array_map('trim', $methods);
-        }
-        $methods = array_intersect($methods, $defaultMethods);
 
-        if (empty($methods)) {
             return $defaultMethods;
         }
 
-        return $methods;
+        return array_intersect($this->method, $defaultMethods);
     }
 
     /**
-     * @inheritdoc
+     * Get hosts.
+     *
+     * @return array|null
      */
-    public function getRoute(): string
+    public function getHosts(): ?array
     {
-        return $this->route;
+        return $this->host;
     }
 
     /**
-     * @inheritdoc
+     * Get priority.
+     *
+     * @return int
      */
-    public function test(string $test): bool
+    public function getPriority(): int
     {
-        return preg_match($this->getRouteRegex(), $test) == 1;
+        return $this->priority;
     }
 
     /**
-     * @inheritdoc
+     * Get options.
      */
-    public function generate(array $parameters)
+    public function getOptions(): array
     {
-        $route = $this->getRoute();
+        return $this->options ?? [];
+    }
 
-        $parametersFound = [];
-        foreach ($this->parameters as $parameter) {
-            if (isset($parameters[$parameter->getName()])) {
-                $value = (string)$parameters[$parameter->getName()];
-            } else {
-                if (!$parameter->hasDefaultValue()) {
-                    return false;
-                }
-
-                $value = (string)$parameter->getDefaultValue();
-            }
-
-            $route = str_replace('{' . $parameter->getName() . '}', $value, $route);
-            $parametersFound[] = $parameter->getName();
+    /**
+     * Get option.
+     *
+     * @param string $name
+     * @param mixed $default
+     *
+     * @return mixed
+     */
+    public function getOption(string $name, mixed $default = null): mixed
+    {
+        if (array_key_exists($name, $this->options)) {
+            return $this->options[$name];
         }
 
-        // Not found parameters
-        $getParameters = [];
-        foreach ($parameters as $parameterName => $parameterValue) {
-            if (!in_array($parameterName, $parametersFound)) {
-                $getParameters[$parameterName] = $parameterValue;
-            }
-        }
-
-        // Construct query string
-        if (!empty($getParameters)) {
-            $getParameters = $this->filterParameters($getParameters);
-            array_walk_recursive(
-                $getParameters,
-                function (&$value) {
-                    $value = (string)$value;
-                }
-            );
-            $httpBuildQuery = http_build_query($getParameters);
-
-            if (!empty($httpBuildQuery)) {
-                $route .= '?' . $httpBuildQuery;
-            }
-        }
-
-        return $route;
+        return $this->parent?->getOption($name, $default) ?? $default;
     }
 
     /**
-     * @inheritdoc
+     * Get context.
+     *
+     * @return mixed
      */
-    public function extractAttributes(string $path): array
+    public function getContext(): mixed
     {
+        return $this->context;
+    }
+
+    /**
+     * Set context.
+     *
+     * @param mixed $context
+     *
+     * @return static
+     */
+    public function setContext(mixed $context): static
+    {
+        $this->context = $context;
+
+        return $this;
+    }
+
+    /**
+     * Is group?
+     *
+     * @return bool
+     */
+    public function isGroup(): bool
+    {
+        return count($this->routes) > 0;
+    }
+
+    /**
+     * Get parent.
+     *
+     * @return Route|null
+     */
+    public function getParent(): ?Route
+    {
+        return $this->parent;
+    }
+
+    /**
+     * Set parent.
+     *
+     * @param Route $parent
+     *
+     * @return static
+     */
+    public function setParent(Route $parent): static
+    {
+        if ($this->parent !== $parent) {
+            $this->path_regex = null;
+        }
+
+        $this->parent = $parent;
+
+        return $this;
+    }
+
+    /**
+     * Test server request.
+     *
+     * @param ServerRequestInterface $request
+     * @param array $attributes
+     *
+     * @return bool
+     */
+    public function test(ServerRequestInterface $request, array &$attributes = []): bool
+    {
+        // Cannot be tested because is a group
+        if ($this->isGroup()) {
+            return false;
+        }
+
         $matches = [];
-        $path = urldecode($path);
-
-        if (preg_match($this->getRouteRegex(), $path, $matches) !== 1) {
-            throw new RoutingException(sprintf('Given path "%s" isn\'t valid for Route "%s"', $path, $this->getName()));
+        if (preg_match('~^' . $this->getPathRegex() . '$~i', $request->getUri()->getPath(), $matches) !== 1) {
+            return false;
         }
 
-        return array_filter($matches, 'is_string', ARRAY_FILTER_USE_KEY);
+        // Accepted method?
+        if (!in_array(strtoupper($request->getMethod()), $this->getMethods())) {
+            return false;
+        }
+
+        // Accepted host?
+        if (null !== $this->getHosts()) {
+            if (!in_array(strtolower($request->getUri()->getHost()), $this->getHosts())) {
+                return false;
+            }
+        }
+
+        $attributes = array_filter($matches, 'is_string', ARRAY_FILTER_USE_KEY);
+
+        return true;
     }
 
     /**
-     * @inheritdoc
-     */
-    public function getNumberOfParameters(): int
-    {
-        return count($this->parameters);
-    }
-
-    /**
-     * Get route with regex replacements.
+     * Generate route.
+     *
+     * @param array $parameters Parameters
+     * @param int $nbUsed
      *
      * @return string
+     * @throws RoutingException
      */
-    public function getRouteRegex(): string
+    public function generate(array $parameters = [], int &$nbUsed = 0): string
     {
-        if (null !== $this->route_regex) {
-            return $this->route_regex;
-        }
+        $this->compile();
 
-        $route = $this;
-        $this->route_regex =
-            '~^' .
-            preg_replace_callback(
-                self::REGEX_PARAMETER,
-                function ($match) use ($route) {
-                    if (isset($route->parameters[$match['name']])) {
-                        return
-                            '(?<' .
-                            $match['name'] .
-                            '>' .
-                            $route->parameters[$match['name']]->getRegexValidation() .
-                            ')';
+        $path =
+            (string)preg_replace_callback(
+                static::REGEX_ATTRIBUTE,
+                function ($matches) use (&$parameters, &$nbUsed) {
+                    if (!array_key_exists($matches['name'], $parameters)) {
+                        $attribute = $this->getAttribute($matches['name']);
+
+                        if (null !== $attribute && $attribute->hasDefault()) {
+                            return $attribute->getDefault();
+                        }
+
+                        return $matches[0];
                     }
 
-                    return $match[0];
-                },
-                $this->route
-            ) .
-            '$~i';
+                    $value = $parameters[$matches['name']];
+                    $nbUsed++;
+                    unset($parameters[$matches['name']]);
 
-        return $this->route_regex;
+                    return (string)$value;
+                },
+                $this->getPath()
+            );
+
+        // Remove optionals attributes
+        $count = 0;
+        do {
+            $path = preg_replace('~\[[^][]*{' . static::REGEX_ATTRIBUTE_NAME . '}[^][]*\]~', '', $path, -1, $count);
+        } while ($count > 0);
+
+        // Check if missing attribute
+        $matches = [];
+        if (preg_match('~{(?<name>' . static::REGEX_ATTRIBUTE_NAME . ')}~', $path, $matches) > 0) {
+            throw RoutingException::missingAttribute($matches['name'], $this->getName());
+        }
+
+        // Remove brackets
+        $count = 0;
+        do {
+            $path = preg_replace('~\[([^][]*)\]~', '\\1', $path, -1, $count);
+        } while ($count > 0);
+
+        // Add over parameters to query string
+        if (count($parameters) > 0) {
+            $parameters = $this->filterParameters($parameters);
+            array_walk_recursive($parameters, fn(&$value) => $value = (string)$value);
+            if (!empty($httpBuildQuery = http_build_query($parameters))) {
+                $path .= '?' . $httpBuildQuery;
+            }
+        }
+
+        return $path;
     }
 
     /**
      * Filter parameters, and remove null parameters.
      *
-     * @param array $params Parameters
+     * @param array $parameters Parameters
      *
      * @return array
      */
-    private function filterParameters(array $params): array
+    private function filterParameters(array $parameters): array
     {
         array_walk(
-            $params,
+            $parameters,
             function (&$param) {
                 if (is_array($param)) {
                     $param = $this->filterParameters($param);
-
-                    $param = array_filter(
-                        $param,
-                        function ($value) {
-                            return $value !== null;
-                        }
-                    );
+                    $param = array_filter($param, fn($value) => null !== $value);
                 }
             }
         );
 
         return array_filter(
-            $params,
+            $parameters,
             function ($value) {
                 if (is_array($value)) {
                     return count($value) > 0;
                 }
 
-                return $value !== null;
+                return null !== $value;
             }
         );
+    }
+
+    /**
+     * Compile.
+     */
+    public function compile(): void
+    {
+        if (null !== $this->path_regex) {
+            return;
+        }
+
+        // Optional parts
+        $path = $this->getPath();
+        $count = 0;
+        do {
+            $path =
+                (string)preg_replace(
+                    '#\[([^\[\]]+)\]#',
+                    '(?:\\1)?',
+                    $path,
+                    -1,
+                    $count
+                );
+        } while ($count > 0);
+
+        $this->path_regex =
+            (string)preg_replace_callback(
+                '/{(?<name>' . Route::REGEX_ATTRIBUTE_NAME . ')}/',
+                function ($match) {
+                    if (null === ($attribute = $this->getAttribute($match['name']))) {
+                        return $match[0];
+                    }
+
+                    return
+                        '(?<' . $attribute->getName() . '>' .
+                        ($attribute->getRegex() ?? static::REGEX_ATTRIBUTE_VALUE) . ')';
+                },
+                $path
+            );
     }
 }
