@@ -14,29 +14,22 @@ declare(strict_types=1);
 
 namespace Berlioz\Router\Exception;
 
-use Exception;
-
 /**
- * Class RoutingException.
+ * Class AmbiguousException.
  *
  * @package Berlioz\Router\Exception
  */
-class RoutingException extends Exception
+class AmbiguousException extends RoutingException
 {
     /**
-     * Missing attribute.
+     * Multiple possible routes.
      *
      * @param string $name
-     * @param string|null $route
      *
      * @return static
      */
-    public static function missingAttribute(string $name, ?string $route): static
+    public static function multiple(string $name)
     {
-        if (null === $route) {
-            return new static(sprintf('Missing attribute "%s" to generate route', $name));
-        }
-
-        return new static(sprintf('Missing attribute "%s" to generate route "%s"', $name, $route));
+        return new static(sprintf('Multiple possible routes named "%s" with given parameters', $name));
     }
 }
