@@ -1,9 +1,9 @@
 <?php
-/**
+/*
  * This file is part of Berlioz framework.
  *
  * @license   https://opensource.org/licenses/MIT MIT License
- * @copyright 2017 Ronan GIRON
+ * @copyright 2020 Ronan GIRON
  * @author    Ronan GIRON <https://github.com/ElGigi>
  *
  * For the full copyright and license information, please view the LICENSE
@@ -14,47 +14,53 @@ declare(strict_types=1);
 
 namespace Berlioz\Router;
 
-use Berlioz\Http\Message\Request;
-use Berlioz\Http\Message\ServerRequest;
-use Berlioz\Http\Message\Stream\PhpInputStream;
-use Berlioz\Http\Message\UploadedFile;
-use Berlioz\Http\Message\Uri;
+use Berlioz\Router\Exception\AmbiguousException;
+use Berlioz\Router\Exception\NotFoundException;
+use Berlioz\Router\Exception\RoutingException;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Log\LoggerAwareTrait;
-
-use function getallheaders;
+use Psr\Log\LoggerInterface;
 
 /**
  * Class Router.
  *
  * @package Berlioz\Router
- * @see \Berlioz\Router\RouterInterface
  */
 class Router implements RouterInterface
 {
     use LoggerAwareTrait;
-
-    /** @var RouteSetInterface Route set */
-    private $routeSet;
-    /** @var ServerRequestInterface Server request */
-    private $serverRequest;
+    use RouteSetTrait;
 
     /**
-     * @inheritdoc
+     * Router constructor.
+     *
+     * @param LoggerInterface|null $logger
      */
-    public function serialize(): string
+    public function __construct(?LoggerInterface $logger = null)
     {
-        return serialize(['routeSet' => $this->routeSet]);
+        if (null !== $logger) {
+            $this->setLogger($logger);
+        }
     }
 
     /**
-     * @inheritdoc
+     * PHP serialize method.
+     *
+     * @return array
      */
-    public function unserialize($serialized)
+    public function __serialize(): array
     {
-        $tmpUnserialized = unserialize($serialized);
+        return ['routes' => $this->routes];
+    }
 
-        $this->routeSet = $tmpUnserialized['routeSet'];
+    /**
+     * PHP unserialize method.
+     *
+     * @param array $data
+     */
+    public function __unserialize(array $data): void
+    {
+        $this->routes = $data['routes'];
     }
 
     /**
@@ -63,195 +69,130 @@ class Router implements RouterInterface
      * @param string $level
      * @param string $message
      */
-    protected function log($level, $message)
+    protected function log(string $level, string $message): void
     {
-        if ($this->logger) {
-            $this->logger->log($level, $message);
-        }
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function getRouteSet(): RouteSetInterface
-    {
-        if (null === $this->routeSet) {
-            $this->routeSet = new RouteSet();
+        if (null === $this->logger) {
+            return;
         }
 
-        return $this->routeSet;
+        $this->logger->log($level, $message);
     }
 
     /**
-     * @inheritdoc
-     */
-    public function setRouteSet(RouteSetInterface $routeSet): RouterInterface
-    {
-        $this->routeSet = $routeSet;
-
-        return $this;
-    }
-
-    /**
-     * Get HTTP headers.
+     * Generate route.
      *
-     * @return array
+     * @param string $name
+     * @param array|RouteAttributes $parameters
+     *
+     * @return string
+     * @throws RoutingException
      */
-    private static function getHeaders(): array
+    public function generate(string $name, array|RouteAttributes $parameters = []): string
     {
-        // Get all headers
-        if (function_exists('\getallheaders')) {
-            return getallheaders() ?: [];
+        $routesFound = [];
+        $exceptions = [];
+        $parameters = $this->generateParameters($parameters);
+
+        /** @var Route $route */
+        foreach ($this->getRoutes($name) as $route) {
+            try {
+                $nbUsed = 0;
+                $path = $route->generate($parameters, $nbUsed);
+                $routesFound[$nbUsed][] = $path;
+            } catch (RoutingException $e) {
+                $exceptions[] = $e;
+            }
         }
 
-        $headers = [];
+        krsort($routesFound);
 
-        $serverVars = $_SERVER;
-        $serverVars['HTTP_CONTENT_TYPE'] = $serverVars['HTTP_CONTENT_TYPE'] ?? $serverVars['CONTENT_TYPE'] ?? null;
-        $serverVars['HTTP_CONTENT_LENGTH'] = $serverVars['HTTP_CONTENT_LENGTH'] ?? $serverVars['CONTENT_LENGTH'] ?? null;
-
-        foreach ($serverVars as $name => $value) {
-            if (substr($name, 0, 5) !== 'HTTP_') {
-                continue;
+        // Get first routes or return false if not found
+        if (false === ($routesFound = reset($routesFound))) {
+            if (!empty($exceptions)) {
+                throw reset($exceptions);
             }
 
-            if (null === $value) {
-                continue;
+            throw new NotFoundException(sprintf('Route "%s" not found with given parameters', $name));
+        }
+
+        if (count($routesFound) > 1) {
+            throw AmbiguousException::multiple($name);
+        }
+
+        return reset($routesFound);
+    }
+
+    private function generateParameters(array|RouteAttributes $parameters = []): array
+    {
+        if ($parameters instanceof RouteAttributes) {
+            return $parameters->routeAttributes();
+        }
+
+        $finalParameters = [];
+
+        array_walk(
+            $parameters,
+            function ($parameter, $key) use (&$finalParameters) {
+                if ($parameter instanceof RouteAttributes) {
+                    $finalParameters = array_merge($finalParameters, $parameter->routeAttributes());
+                    return;
+                }
+
+                if (is_array($parameter)) {
+                    $finalParameters = array_merge($finalParameters, $parameter);
+                    return;
+                }
+
+                $finalParameters = array_merge($finalParameters, [$key => $parameter]);
             }
+        );
 
-            $headers[str_replace(
-                ' ',
-                '-',
-                ucwords(strtolower(str_replace('_', ' ', substr($name, 5))))
-            )] = $value;
-        }
-
-        return $headers;
+        return $finalParameters;
     }
 
     /**
-     * Make server request.
+     * Is valid request?
      *
-     * @return ServerRequestInterface
+     * @param ServerRequestInterface $request
+     *
+     * @return bool
      */
-    public static function makeServerRequest(): ServerRequestInterface
+    public function isValid(ServerRequestInterface $request): bool
     {
-        // HTTP Method
-        if (!empty($_SERVER['HTTP_X_HTTP_METHOD_OVERRIDE'])) {
-            $method = mb_strtoupper($_SERVER['HTTP_X_HTTP_METHOD_OVERRIDE']);
-        } elseif (isset($_SERVER['REQUEST_METHOD'])) {
-            $method = mb_strtoupper($_SERVER['REQUEST_METHOD']);
-        } else {
-            $method = 'GET';
+        /** @var Route $route */
+        foreach ($this->getRoutes() as $route) {
+            if ($route->test($request)) {
+                return true;
+            }
         }
 
-        // Path
-        $path = null;
-        if (isset($_SERVER['REQUEST_URI'])) {
-            $path = parse_url($_SERVER['REQUEST_URI'], PHP_URL_PATH);
-        }
-
-        // Query string
-        $queryString = $_SERVER['REDIRECT_QUERY_STRING'] ?? $_SERVER['QUERY_STRING'] ?? '';
-
-        // Request URI
-        $requestUri = new Uri(
-            $_SERVER['REQUEST_SCHEME'] ?? '',
-            $_SERVER['HTTP_HOST'] ?? '',
-            $_SERVER['SERVER_PORT'] ?? 80,
-            $path,
-            $queryString,
-            '',
-            $_SERVER['PHP_AUTH_USER'] ?? '',
-            $_SERVER['PHP_AUTH_PW'] ?? ''
-        );
-
-        // Server request
-        return new ServerRequest(
-            $method,
-            $requestUri,
-            static::getHeaders(),
-            $_COOKIE,
-            $_SERVER,
-            new PhpInputStream(),
-            UploadedFile::parseUploadedFiles($_FILES)
-        );
+        return false;
     }
 
     /**
-     * @inheritdoc
+     * Handle server request.
+     *
+     * @param ServerRequestInterface $request
+     *
+     * @return RouteInterface|null
      */
-    public function getServerRequest(): ServerRequestInterface
-    {
-        if (null === $this->serverRequest) {
-            $this->serverRequest = static::makeServerRequest();
-        }
-
-        return $this->serverRequest;
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function setServerRequest(ServerRequestInterface $serverRequest): RouterInterface
-    {
-        $this->serverRequest = $serverRequest;
-
-        return $this;
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function generate(string $name, array $parameters = [])
-    {
-        if (null === ($route = $this->getRouteSet()->getByName($name))) {
-            return false;
-        }
-
-        return $route->generate($parameters);
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function isValid(string $path, string $method = Request::HTTP_METHOD_GET): bool
-    {
-        $uri = Uri::createFromString($path);
-
-        return null !== ($this->getRouteSet()->searchRoute($uri, $method));
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function handle(?ServerRequestInterface &$serverRequest = null): ?RouteInterface
+    public function handle(ServerRequestInterface &$request): ?RouteInterface
     {
         // Log
         $this->log('debug', sprintf('%s', __METHOD__));
 
-        // Server request
-        if (null === $serverRequest) {
-            $serverRequest = $this->getServerRequest();
+        $attributes = [];
+        $route = $this->searchRoute($request, $attributes);
 
+        if (null !== $route) {
             // Log
-            $this->log('debug', sprintf('%s / ServerRequest created', __METHOD__));
+            $this->log('debug', sprintf('%s / Route found', __METHOD__));
+
+            // Add attributes to server request
+            foreach ($attributes as $name => $value) {
+                $request = $request->withAttribute($name, $value);
+            }
         }
-
-        $route = $this->getRouteSet()->searchRoute($serverRequest->getUri(), $serverRequest->getMethod());
-        if (null === $route) {
-            return null;
-        }
-
-        // Log
-        $this->log('debug', sprintf('%s / Route found', __METHOD__));
-
-        // Attributes
-        foreach ($route->extractAttributes($serverRequest->getUri()->getPath()) as $name => $value) {
-            $serverRequest = $serverRequest->withAttribute($name, $value);
-        }
-
-        $this->setServerRequest($serverRequest);
 
         // Log
         $this->log('debug', sprintf('%s / ServerRequest completed', __METHOD__));
